+++ src/main/java/frc/robot/Robot.java
@@ -10,6 +10,8 @@ package frc.robot;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
+import frc.robot.misc.ControlChooser;
+import frc.robot.misc.SmartDashboardInterface;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
