+++ src/main/java/frc/robot/misc/ControlChooser.java
@@ -2,7 +2,7 @@
  * @author Hunter Pugh
  */
 
-package frc.robot;
+package frc.robot.misc;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.button.Button;
